定义仓储接口和实现、定义工作单元接口和实现

+++ MMS.Data/BaseRepository.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Data.Entity;
+using System.Data.Entity.Infrastructure;
 using System.Linq;
 using System.Linq.Expressions;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace MMS.Data
 {
@@ -15,7 +14,7 @@ namespace MMS.Data
         public BaseRepository(DatabaseContext context)
         {
             _context = context;
-            _dbSet = _context.Set<TEntity>();
+            _dbSet = context.Set<TEntity>();
         }
 
         #region 单体查询
@@ -26,12 +25,33 @@ namespace MMS.Data
         {
             return _dbSet.Find(id);
         }
+
         /// <summary>
         /// 根据条件查询单个实体
         /// </summary>
         public TEntity Find(Expression<Func<TEntity, bool>> predicate)
         {
             return _dbSet.FirstOrDefault(predicate);
+        }
+
+        /// <summary>
+        /// 根据条件查询单个实体(包括导航属性)
+        /// </summary>
+        /// <param name="predicate">查询条件</param>
+        /// <param name="includeProperties">要加载的导航属性</param>
+        /// <returns></returns>
+        public TEntity Find(Expression<Func<TEntity, bool>> predicate, string[] includeProperties)
+        {
+            if (includeProperties.Length<=0)
+            {
+                Find(predicate);
+            }
+            IQueryable<TEntity> query = _dbSet;
+            foreach (string property in includeProperties)
+            {
+                query = query.Include(property);
+            }
+            return query.FirstOrDefault(predicate);
         }
         #endregion
 
@@ -53,78 +73,94 @@ namespace MMS.Data
         }
 
         /// <summary>
-        /// 查询排序实体列表
+        /// 按条件查询的分页列表
         /// </summary>
-        /// <typeparam name="TOrder">排序列类型</typeparam>
-        /// <param name="order">排序条件</param>
-        /// <param name="asc">是否升序排列</param>
-        public IQueryable<TEntity> FindList<TOrder>(Expression<Func<TEntity, TOrder>> order, bool asc)
-        {
-            return asc ? _dbSet.OrderBy(order):_dbSet.OrderByDescending(order);
-        }
-
-        /// <summary>
-        /// 查询排序实体列表并取指定数量的实体
-        /// </summary>
-        /// <typeparam name="TOrder">排序列类型</typeparam>
-        /// <param name="order">排序条件</param>
-        /// <param name="asc">是否升序排列</param>
-        /// <param name="number">获取实体数量</param>
-        /// <returns></returns>
-        public IQueryable<TEntity> FindList<TOrder>(Expression<Func<TEntity, TOrder>> order, bool asc, int number)
-        {
-            return asc ? _dbSet.OrderBy(order).Take(number) : _dbSet.OrderByDescending(order).Take(number);
-        }
-
-        /// <summary>
-        /// 按条件查询排序实体列表
-        /// </summary>
-        /// <typeparam name="TOrder">排序列类型</typeparam>
         /// <param name="predicate">查询条件</param>
-        /// <param name="order">排序条件</param>
-        /// <param name="asc">是否升序排列</param>
-        /// <returns></returns>
-        public IQueryable<TEntity> FindList<TOrder>(Expression<Func<TEntity, bool>> predicate, Expression<Func<TEntity, TOrder>> order, bool asc)
-        {
-            return asc ? _dbSet.Where(predicate).OrderBy(order) : _dbSet.Where(predicate).OrderByDescending(order);
-        }
-
-        /// <summary>
-        /// 按条件查询排序实体列表并获取指定数量的实体
-        /// </summary>
-        /// <typeparam name="TOrder">排序键类型</typeparam>
-        /// <param name="predicate">查询条件</param>
-        /// <param name="order">排序条件</param>
-        /// <param name="asc">是否升序排列</param>
-        /// <param name="number">获取实体数量</param>
-        /// <returns></returns>
-        public IQueryable<TEntity> FindList<TOrder>(Expression<Func<TEntity, bool>> predicate, Expression<Func<TEntity, TOrder>> order, bool asc, int number)
-        {
-            return asc ? _dbSet.Where(predicate).OrderBy(order).Take(number) : _dbSet.Where(predicate).OrderByDescending(order).Take(number);
-        }
-
-        /// <summary>
-        /// 按条件查询排列实体列表并获取分页后的实体
-        /// </summary>
-        /// <typeparam name="TOrder">排序列类型</typeparam>
-        /// <param name="predicate">查询条件</param>
-        /// <param name="order">排序条件</param>
-        /// <param name="asc">是否升序排列</param>
         /// <param name="pageSize">每页记录数量</param>
         /// <param name="pageIndex">当前页码</param>
         /// <param name="totalCount">总记录数量</param>
-        /// <returns></returns>
-        public IQueryable<TEntity> FindPageList<TOrder>(Expression<Func<TEntity, bool>> predicate, Expression<Func<TEntity, TOrder>> order, bool asc, int pageSize, int pageIndex, out int totalCount)
+        public IQueryable<TEntity> FindPageList(Expression<Func<TEntity, bool>> predicate, int pageSize, int pageIndex, out int totalCount)
         {
+            IQueryable<TEntity> query = _dbSet.Where(predicate);
+            totalCount = query.Count();
             pageSize = pageSize <= 0 ? 20 : pageSize;
-            pageIndex= pageIndex <= 0 ? 1 : pageIndex;
-            totalCount = _dbSet.Where(predicate).Count();
-            return asc ? _dbSet.Where(predicate).OrderBy(order).Skip(pageSize * (pageIndex - 1)).Take(pageSize) : _dbSet.Where(predicate).OrderByDescending(order).Skip(pageSize * (pageIndex - 1)).Take(pageSize);
+            pageIndex = pageIndex <= 0 ? 1 : pageIndex;
+            return query.Skip(pageSize * (pageIndex - 1)).Take(pageSize) ;
         }
 
-        public IQueryable<TEntity> FindPageList<TOrder>(Expression<Func<TEntity, bool>> predicate, Expression<Func<TEntity, TOrder>> order, bool asc, int pageSize, int pageIndex, out int totalCount, string[] includeProperties)
+        /// <summary>
+        /// 按条件查询的分页列表(包括导航属性)
+        /// </summary>
+        /// <param name="predicate">查询条件</param>
+        /// <param name="pageSize">每页记录数量</param>
+        /// <param name="pageIndex">当前页码</param>
+        /// <param name="totalCount">总记录数量</param>
+        /// <param name="includeProperties">要加载的导航属性名(列名)</param>
+        /// <returns></returns>
+        public IQueryable<TEntity> FindPageList(Expression<Func<TEntity, bool>> predicate, int pageSize, int pageIndex, out int totalCount, string[] includeProperties)
         {
-            throw new NotImplementedException();
+            if (includeProperties.Length <= 0)
+            {
+                FindPageList(predicate, pageSize, pageIndex, out totalCount);
+            }
+            IQueryable<TEntity> query = _dbSet;
+            foreach (string property in includeProperties)
+            {
+                query = query.Include(property);
+            }
+            query = query.Where(predicate);
+            totalCount = query.Count();
+            pageSize = pageSize <= 0 ? 20 : pageSize;
+            pageIndex = pageIndex <= 0 ? 1 : pageIndex;
+            return query.Skip(pageSize * (pageIndex - 1)).Take(pageSize);
+        }
+
+        /// <summary>
+        /// 获取按条件查询的分页排序列表
+        /// </summary>
+        /// <typeparam name="TOrder">排序列类型</typeparam>
+        /// <param name="predicate">查询条件</param>
+        /// <param name="order">排序条件</param>
+        /// <param name="isAsc">是否升序排列</param>
+        /// <param name="pageSize">每页记录数量</param>
+        /// <param name="pageIndex">当前页码</param>
+        /// <param name="totalCount">总记录数量</param>
+        public IQueryable<TEntity> FindPageList<TOrder>(Expression<Func<TEntity, bool>> predicate, Expression<Func<TEntity, TOrder>> order, bool isAsc, int pageSize, int pageIndex, out int totalCount)
+        {
+            IQueryable<TEntity> query = _dbSet.Where(predicate);
+            pageSize = pageSize <= 0 ? 20 : pageSize;
+            pageIndex = pageIndex <= 0 ? 1 : pageIndex;
+            totalCount = query.Count();
+            return isAsc ? query.OrderBy(order).Skip(pageSize * (pageIndex - 1)).Take(pageSize) : query.OrderByDescending(order).Skip(pageSize * (pageIndex - 1)).Take(pageSize);
+        }
+
+        /// <summary>
+        /// 获取按条件查询的分页排序列表(包括导航属性)
+        /// </summary>
+        /// <typeparam name="TOrder">排序列类型</typeparam>
+        /// <param name="predicate">查询条件</param>
+        /// <param name="order">排序条件</param>
+        /// <param name="isAsc">是否升序排列</param>
+        /// <param name="pageSize">每页记录数量</param>
+        /// <param name="pageIndex">当前页码</param>
+        /// <param name="totalCount">总记录数量</param>
+        /// <param name="includeProperties">要加载的导航属性名(列名)</param>
+        public IQueryable<TEntity> FindPageList<TOrder>(Expression<Func<TEntity, bool>> predicate, Expression<Func<TEntity, TOrder>> order, bool isAsc, int pageSize, int pageIndex, out int totalCount, string[] includeProperties)
+        {
+            if (includeProperties.Length<=0)
+            {
+                FindPageList(predicate, order, isAsc, pageSize, pageIndex, out totalCount);
+            }
+            IQueryable<TEntity> query = _dbSet;
+            foreach (string property in includeProperties)
+            {
+                query = query.Include(property);
+            }
+            query = query.Where(predicate);
+            totalCount = query.Count();
+            pageSize = pageSize <= 0 ? 20 : pageSize;
+            pageIndex = pageIndex <= 0 ? 1 : pageIndex;
+            return isAsc?query.OrderBy(order).Skip(pageSize * (pageIndex - 1)).Take(pageSize) : query.OrderByDescending(order).Skip(pageSize * (pageIndex - 1)).Take(pageSize);
         }
         #endregion
 
@@ -134,6 +170,10 @@ namespace MMS.Data
         /// </summary>
         public void Add(TEntity entity)
         {
+            if (entity == null)
+            {
+                return;
+            }
             _dbSet.Add(entity);
         }
         /// <summary>
@@ -141,6 +181,10 @@ namespace MMS.Data
         /// </summary>
         public void Add(IEnumerable<TEntity> entities)
         {
+            if (entities.Count() <= 0 || entities == null)
+            {
+                return;
+            }
             //_context.Configuration.AutoDetectChangesEnabled = false;
             _dbSet.AddRange(entities);
         }
@@ -152,6 +196,10 @@ namespace MMS.Data
         /// </summary>
         public void Delete(TEntity entity)
         {
+            if (entity == null)
+            {
+                return;
+            }
             if (_context.Entry(entity).State == EntityState.Detached)
             {
                 _dbSet.Attach(entity);
@@ -164,10 +212,7 @@ namespace MMS.Data
         public void Delete(object id)
         {
             TEntity entity = _dbSet.Find(id);
-            if (entity != null)
-            {
-                Delete(entity);
-            }
+            Delete(entity);
         }
         /// <summary>
         /// 批量删除实体
@@ -175,6 +220,10 @@ namespace MMS.Data
         /// <param name="entities"></param>
         public void Delete(IEnumerable<TEntity> entities)
         {
+            if (entities.Count() <= 0 || entities == null)
+            {
+                return;
+            }
             //TODO:是否需要先设置实体状态？关联到上下文？
             _dbSet.RemoveRange(entities);
         }
@@ -184,7 +233,7 @@ namespace MMS.Data
         public void Delete(Expression<Func<TEntity, bool>> predicate)
         {
             var entities = _dbSet.Where(predicate);
-            if (entities.Count() <= 0)
+            if (entities.Count() <= 0 || entities == null)
             {
                 return;
             }
@@ -200,14 +249,46 @@ namespace MMS.Data
         #endregion
 
         #region 编辑
+        /// <summary>
+        /// 更新实体
+        /// </summary>
         public void Edit(TEntity entity)
         {
+            if (entity == null)
+            {
+                return;
+            }
             if (_context.Entry(entity).State == EntityState.Detached)
             {
                 _dbSet.Attach(entity);
             }
             _context.Entry(entity).State = EntityState.Modified;
         }
+        /// <summary>
+        /// 按指定属性(列)更新实体
+        /// </summary>
+        /// <param name="changedProperties">需要更新的属性名（列名）集合</param>
+        public void Edit(TEntity entity, string[] changedProperties)
+        {
+            if (entity == null)
+            {
+                return;
+            }
+            if (changedProperties.Length <= 0)
+            {
+                Edit(entity);
+                return;
+            }
+            DbEntityEntry entry = _context.Entry(entity);
+            entry.State = EntityState.Unchanged;
+            foreach (string property in changedProperties)
+            {
+                entry.Property(property).IsModified = true;
+            }
+            //关闭EF对于实体的合法性验证参数
+            // db.Configuration.ValidateOnSaveEnabled = false;
+        }
+
         #endregion
 
     }
